# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Cryptocoin Exodo",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "https://criptocoin.rj.r.appspot.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/users": {
      "get": {
        "tags": [
          "Usuario"
        ],
        "description": "Retorna todos os usuarios, somente admin",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          },
          "500": {
            "description": "Erro Interno do Servidor"
          }
        }
      }
    },
...
    "/taxaSaque": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "description": "Define a taxa do saque",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "taxa": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
    "/deleteUser/{id}": {
      "post": {
        "tags": [
          "Usuario"
        ],
        "description": "Deleta um usuario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Não autorizado"
          }
        }
      }
    },
...
  },
...
}